Build with -Wall -Wextra -Werror; fix what that surfaced
Warnings are errors now for everything hand-written. The generated scanner
and parser are exempt from -Werror: their warnings belong to flex and bison
and change with the tool version.

Turning it on found a real bug. save_array_to_ihex() still ended with
"return -1", so "-x ihex" reported "Failed to write Intel HEX output" and
exited 1 on every successful run - the file itself was correct all along.
This is the item #32 listed and that an earlier commit claimed to have
fixed; only half of that edit actually landed, and nothing caught it
because the ihex path had no test.

Also fixed, from -Wextra:
- four fwrite() results in ihex.c and one in tap.c were compared against
  an int, mixing signedness
- the opcode table terminator {0, 0} left later fields implicit

New test ihex_output assembles the ROM to Intel HEX and checks the records
parse, every checksum is right, and the decoded bytes match the reference
image. Verified it fails if the return value is broken again.

ctest: 7/7 pass, clean build with no warnings.

+++ src/asmdasm/tap.c
@@ -209,7 +209,7 @@ int tap_create(struct t_tap_info *p_tap, FILE *out)
     l += DATA_HDR_OFFSET;
 
     /* write all headers to file */
-    if (fwrite(p, 1, l, out) != l)
+    if (fwrite(p, 1, l, out) != (size_t)l)
     {
         return -1;
     }
